Lots of updates on this one.
We have our very own bitmap vectorizer!!! It works pretty well, but for the moment it vectorizes on load, so it's a bit slow to start up. I'll probably need to move that to an offline process.

Sprites are now organized like normal. But if you try to do that with the XCode editor, they mess up something internally which causes the sketch outlines to not line up with the sprite. If you delete the sprite and re-add it, it works until you add another sprite, so it's clearly a bug in XCode. :( So for now, the sprites are added to the scene hard-coded.

In order for the system to know what bitmap was used to load a sprite, you have to store the texture name inside the sprite's name field. (you can't query a sprite for it's texture image.)

Fixed a number of issues with view space inconsistencies (the background didn't quite cover the screen, things had to be flipped or the would be rendered upside down, and a few other issues that I forget.

Added temp artwork.

Cleaned up the extensions for CGVector, CGPoint and added a new Point2D which stores integer X/Y coordinates. Not sure if I ever ended up using that or not, though.