Added a simple progress bar art.
Unfortunately, since our progress just used a mask to unveil the progress indicator as the level was loaded, we don't convey progress to the user because they're not given a 'start' and an 'end'. So I had to re-work the progress a bit. I just added steve running along a sketched line.